Support threads
To support threads, we the create a new user/channel with the thread
timestamp as a suffix, then use that to separate where the messages are
displayed. Each thread will show up as a new user/channel in the IRC
interface.

To avoid getting notifications for every message on every thread, we
only add to the channel list those threads that the current user is part
of.

For replying to a thread, we reverse that by splitting the thread
timestamp from the channel/user name and modifying the POST request.