Load conversation history on connect
Once we are successfully connected and IRC is listening for events, we
can query the Slack API for all unread conversation history and send
them as new messages to their corresponding channels. We prepend the
original date of the message to distinguish them since the IRC client
will treat them as brand new messages.